(#39) spellchecker: extract to a separate optional module
