+++ core/src/main/scala/me/fornever/haskeletor/core/HaskellLanguage.java
@@ -6,7 +6,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package me.fornever.haskeletor;
+package me.fornever.haskeletor.core;
 
 import com.intellij.lang.Language;
 import org.jetbrains.annotations.NotNull;
