+++ spellchecker/src/main/resources/dictionary/haskell.dic
