+++ spellchecker/src/main/scala/me/fornever/haskeletor/spellchecker/HaskellBundledDictionaryProvider.scala
