+++ src/main/scala/me/fornever/haskeletor/cabal/CabalFileType.java
@@ -10,6 +10,7 @@ package me.fornever.haskeletor.cabal;
 
 import com.intellij.lang.Language;
 import com.intellij.openapi.fileTypes.LanguageFileType;
+import me.fornever.haskeletor.core.cabal.CabalLanguage;
 import me.fornever.haskeletor.icons.HaskellIcons;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
