+++ src/main/scala/me/fornever/haskeletor/cabal/lang/psi/CabalElementType.java
@@ -9,7 +9,7 @@
 package me.fornever.haskeletor.cabal.lang.psi;
 
 import com.intellij.psi.tree.IElementType;
-import me.fornever.haskeletor.cabal.CabalLanguage;
+import me.fornever.haskeletor.core.cabal.CabalLanguage;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
